refactor(App): split files (#52573)
